Moved functions to separate files
